add shouldDelete to repair

+++ satellite/repair/repairer/repairer.go
@@ -110,7 +110,7 @@ func (service *Service) worker(ctx context.Context, seg *pb.InjuredSegment) (err
 	workerStartTime := time.Now().UTC()
 
 	service.log.Info("Limiter running repair on segment", zap.Binary("segment", seg.GetPath()))
-	err = service.repairer.Repair(ctx, string(seg.GetPath()))
+	shouldDelete, err := service.repairer.Repair(ctx, string(seg.GetPath()))
 	if err != nil {
 		if IrreparableError.Has(err) {
 			service.log.Error(
@@ -122,11 +122,12 @@ func (service *Service) worker(ctx context.Context, seg *pb.InjuredSegment) (err
 
 		return Error.New("repairing injured segment: %v", err)
 	}
-
-	service.log.Info("Deleting segment from repair queue", zap.Binary("segment", seg.GetPath()))
-	err = service.queue.Delete(ctx, seg)
-	if err != nil {
-		return Error.New("deleting repaired segment from the queue: %v", err)
+	if shouldDelete {
+		service.log.Info("Deleting segment from repair queue", zap.Binary("segment", seg.GetPath()))
+		err = service.queue.Delete(ctx, seg)
+		if err != nil {
+			return Error.New("deleting repaired segment from the queue: %v", err)
+		}
 	}
 
 	repairedTime := time.Now().UTC()
